refactor(PaymentClient): PaymentClient 를 도메인 패키지로 이동
