+++ src/test/java/roomescape/TestPaymentConfig.java
@@ -2,8 +2,8 @@ package roomescape;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import roomescape.service.FakePayment;
-import roomescape.service.PaymentClient;
+import roomescape.domain.FakePayment;
+import roomescape.domain.PaymentClient;
 
 @TestConfiguration
 public class TestPaymentConfig {
