+++ src/test/java/roomescape/domain/FakePayment.java
@@ -1,8 +1,7 @@
-package roomescape.service;
+package roomescape.domain;
 
 import static roomescape.api.TossPaymentExceptionType.INVALID_REQUEST;
 
-import roomescape.domain.Payment;
 import roomescape.domain.Payment.State;
 import roomescape.dto.PaymentRequest;
 import roomescape.exception.PaymentException;
